Refactor CloudFormation
